limavm: Support `rdd limavm create --dry-run`
This adds the option to use (server-side) dry-run to create lima vms with
full validation.  It actually still creates the config map as needed, as
that is required to pass validation; the config map is deleted after if it
is a dry-run.

We may want to consider embedding the configuration into the LimaVM object
instead in the future, so that this would not be an issue.

Signed-off-by: Mark Yen <mark.yen@suse.com>