rdd: implement `rdd svc log`
This new command prints out the last known logs of the corresponding RDD
instance.  By default, stderr logs are printed; `rdd svc log --stdout` can
be used to print the stdout logs instead.  `rdd svc log --follow` can be
used to tail the logs until the user interrupts it.

We are using the pulumi fork of `github.com/nxadm/tail` because that has
some fixes around stopping at the end of the file that would be important
for our use case (when not using `--follow`) that has pull requests
upstream that have not been merged.

Signed-off-by: Mark Yen <mark.yen@suse.com>