Allow multiple controller managers
This makes it possible to support multiple shared controller managers (each
in a different process).  This is to make it easier to reuse the code; we
can have multiple processes running different controllers in the future to
e.g. allow testing-only controllers to be used with the normal build.

The config map is changed so that the key is the name of the controller
process (with the default one named "embedded"), and the value is JSON-
serialized `ControllerManagerInfo`.  JSON is chosen so that we can dig into
it as part of the BATS tests.  We also no longer track the instance in the
info, because that is fixed per Kubernetes API controller so it does not
actually provide any useful information.

To test:
- Build all the executable (with `make build`)
- Run the executable with a subset of controllers:
  `bin/rdd --controllers rdd,lima`
- Run the app controller externally:
  `bin/app-controller`

Everything should work as before, the only difference being multiple
controller processes co-existing.

Signed-off-by: Mark Yen <mark.yen@suse.com>